+++ app/access-control/controllers/ldap-controller.html
@@ -1,7 +1,7 @@
 <loader loading="loading"></loader>
 <div class="ldap" id="configuration-ldap">
   <div class="row column">
-    <h1>LDAP role group settings</h1>
+    <h1>LDAP</h1>
   </div>
   <div class="row column">
     <p>Configure LDAP settings and manage role groups.</p>
@@ -46,7 +46,7 @@
               <span>SSL certificates must be uploaded to secure LDAP using SSL.</span>
             </div>
             <div class="ldap__certificate-info">
-              <a href="#/configuration/certificate">Go to SSL certificates</a>
+              <a href="#/access-control/ssl-certificates">Go to SSL certificates</a>
             </div>
           </div>
           <div class="large-9 columns ldap__server-info">
