+++ app/access-control/controllers/user-accounts-modal-settings.html
