+++ app/access-control/controllers/user-controller.js
@@ -1,15 +1,15 @@
 /**
  * Controller for user Accounts
  *
- * @module app/users
- * @exports userAccountsController
- * @name userAccountsController
+ * @module app/access-control
+ * @exports userController
+ * @name userController
  */
 
 window.angular && (function(angular) {
   'use strict';
 
-  angular.module('app.users').controller('userAccountsController', [
+  angular.module('app.accessControl').controller('userController', [
     '$scope', 'APIUtils', 'toastService', '$uibModal', '$q',
     function($scope, APIUtils, toastService, $uibModal, $q) {
       $scope.loading;
