+++ app/access-control/directives/role-table.js
@@ -5,7 +5,7 @@ window.angular && (function(angular) {
    * Role table
    * Table of privilege role descriptions
    */
-  angular.module('app.users').directive('roleTable', [
+  angular.module('app.accessControl').directive('roleTable', [
     '$sce',
     function($sce) {
       return {
