+++ app/access-control/styles/certificate.scss
