+++ app/access-control/styles/index.scss
@@ -1,0 +1,3 @@
+@import "./user-accounts.scss";
+@import "./certificate.scss";
+@import "./ldap.scss";
