+++ app/configuration/styles/index.scss
@@ -2,5 +2,3 @@
 @import "./snmp.scss";
 @import "./date-time.scss";
 @import "./firmware.scss";
-@import "./certificate.scss";
-@import "./ldap.scss";
