+++ app/styles/index.scss
@@ -6,5 +6,5 @@ $charset: "UTF-8";
 @import "../server-control/styles/index.scss";
 @import "../server-health/styles/index.scss";
 @import "../configuration/styles/index.scss";
-@import "../users/styles/index.scss";
+@import "../access-control/styles/index.scss";
 @import "../redfish/styles/index.scss";
