--- app/users/styles/index.scss
@@ -1,1 +1,0 @@
-@import "./user-accounts.scss";
